+++ packages/plugin-button/web/src/components/button-component.jsx
@@ -14,7 +14,7 @@ class ButtonComponent extends PureComponent {
   render() {
     const {
       componentData: { button },
-      settings: { onClickEvent },
+      settings: { onClick },
     } = this.props;
     const { theme } = this.props;
     const buttonText = button.settings.buttonText;
@@ -34,7 +34,7 @@ class ButtonComponent extends PureComponent {
         rel={rel}
         buttonText={buttonText}
         theme={theme}
-        onClickEvent={onClickEvent}
+        onClick={onClick}
       />
     );
   }