+++ packages/plugin-button/web/src/components/button-viewer.jsx
@@ -5,37 +5,24 @@ import { mergeStyles } from 'wix-rich-content-common';
 import styles from '../../statics/styles/default-styles.scss';
 class ButtonViewer extends PureComponent {
   render() {
-    const { theme, onClickEvent } = this.props;
+    const { theme, onClick } = this.props;
     this.styles = this.styles || mergeStyles({ styles, theme });
     const { url, style, target, rel, buttonText } = this.props;
-    const isActionButton = Boolean(onClickEvent);
-
-    if (isActionButton) {
-      return (
-        // eslint-disable-next-line jsx-a11y/click-events-have-key-events
-        <div
-          className={this.styles.button_container}
-          style={style}
-          target={target}
-          rel={rel}
-          onClick={onClickEvent}
-        >
-          <div className={this.styles.button_text}>{buttonText}</div>
-        </div>
-      );
-    }
+    const isActionButton = Boolean(onClick);
+    const Component = isActionButton ? 'div' : 'a';
+    let props = { className: this.styles.button_container, style };
+    props = isActionButton
+      ? { onClick, ...props }
+      : {
+          href: url,
+          target,
+          rel,
+          ...props,
+        };
     return (
-      <div>
-        <a
-          className={this.styles.button_container}
-          href={url}
-          style={style}
-          target={target}
-          rel={rel}
-        >
-          <div className={this.styles.button_text}>{buttonText}</div>
-        </a>
-      </div>
+      <Component {...props}>
+        <div className={this.styles.button_text}>{buttonText}</div>
+      </Component>
     );
   }
 }
@@ -47,7 +34,7 @@ ButtonViewer.propTypes = {
   rel: PropTypes.string,
   buttonText: PropTypes.string,
   theme: PropTypes.object.isRequired,
-  onClickEvent: PropTypes.func,
+  onClick: PropTypes.func,
 };
 
 export default ButtonViewer;