+++ packages/plugin-button/web/src/toolbar/buttonInputModal.jsx
@@ -32,7 +32,7 @@ export default class ButtonInputModal extends Component {
       initialComponentData: { ...button },
       isHover: false,
       activeTab: settingsTabValue,
-      isLinkNeeded: !this.props.settings.onClickEvent,
+      shouldShowLink: !this.props.settings.onClick,
     };
 
     this.setScrollbarRef = element => {
@@ -72,11 +72,11 @@ export default class ButtonInputModal extends Component {
   };
 
   onConfirm = () => {
-    const { validUrl, isLinkNeeded } = this.state;
+    const { validUrl, shouldShowLink } = this.state;
     const {
       helpers: { closeModal },
     } = this.props;
-    if (!isLinkNeeded || validUrl) {
+    if (!shouldShowLink || validUrl) {
       this.setState({ submitted: true, isOpen: false });
       closeModal();
     } else {
@@ -126,14 +126,14 @@ export default class ButtonInputModal extends Component {
 
   render() {
     const { theme, t, uiSettings, doneLabel, cancelLabel, isMobile } = this.props;
-    const { isLinkNeeded } = this.state;
+    const { shouldShowLink } = this.state;
     const { styles } = this;
     const settingTabLabel = (
       <div className={styles.button_inputModal_settingTab}>
         <div className={styles.button_inputModal_tabTitle}>
           <p className={styles.button_inputModal_tabLabel}>{t('ButtonModal_Settings_Tab')}</p>
         </div>
-        {isLinkNeeded && (
+        {shouldShowLink && (
           <div className={styles.button_inputModal_errorIcon}>
             {!this.state.validUrl ? <ErrorIcon width="18" height="18" /> : null}
           </div>
@@ -157,7 +157,7 @@ export default class ButtonInputModal extends Component {
         linkInputRef={ref => {
           this.linkInput = ref;
         }}
-        isLinkNeeded={isLinkNeeded}
+        shouldShowLink={shouldShowLink}
       />
     );
     const designComponent = (