+++ CHANGELOG.md
@@ -17,7 +17,7 @@
   
 ### :rocket: New Feature
 - `button`
-  - [#953](https://github.com/wix-incubator/rich-content/pull/953) action button: by passing onClickEvent callback, on click triggers it
+  - [#953](https://github.com/wix-incubator/rich-content/pull/953) action button: by passing onClick callback, on click triggers it
 </details>
 <hr/>